Moved swipe card to home page
